added ignoring _local_ prefix and added slight protection to build.sh

+++ build.sh
@@ -5,4 +5,6 @@ cmake -G Ninja -B build \
   -DCMAKE_BUILD_TYPE=RelWithDebInfo \
   -DSFML_BUILD_NETWORK=OFF -DSFML_BUILD_AUDIO=OFF .
 cmake --build build/
-ln -sr resources build/bin/resources
+if [[ -e "build/bin/resources" ]]; then
+  ln -sr resources build/bin/resources
+fi
